[vim] tweak CloseBufferAndDisplayNext function
The function previously would switch to the next buffer without losing
the split window. However, sometimes you move to a buffer that happens
to be open in another window; if you close that buffer (which is open in
more than one window) using bdelete, _both_ buffers are closed, and the
split is gone.

Using :hide instead of :bdelete avoids that problem.

Signed-off-by: Marco Magdy <mmagdy@gmail.com>